# Proof of concept: speed up plugin navigation