--- a/resources/lib/classes.py
+++ b/resources/lib/classes.py
@@ -1,14 +1,30 @@
 import urlparse
 import urllib
 import unicodedata
+import json
+import datetime
+import xbmcgui
+
+from collections import deque
+
+from aussieaddonscommon import session
+from aussieaddonscommon import utils
 
 
 class genre(object):
-    def __init__(self):
+    def __init__(self, attrs=None):
+        if attrs:
+            for attr, val in attrs.items():
+                setattr(self, attr, val)
+            return
+
         self.fanart = None
         self.thumb = None
         self.title = None
         self.genre_slug = None
+
+    def __repr__(self):
+        return type(self).__name__ + '(' + str(dict(vars(self).items())) + ')'
 
     def make_kodi_url(self):
         d = vars(self)
@@ -25,7 +41,12 @@ class genre(object):
 
 
 class series(object):
-    def __init__(self):
+    def __init__(self, attrs=None):
+        if attrs:
+            for attr, val in attrs.items():
+                setattr(self, attr, val)
+            return
+
         self.multi_season = None
         self.fanart = None
         self.thumb = None
@@ -35,6 +56,10 @@ class series(object):
         self.season_slug = None
         self.title = None
         self.genre = None
+        self.desc = None
+
+    def __repr__(self):
+        return type(self).__name__ + '(' + str(dict(vars(self).items())) + ')'
 
     def get_title(self):
         if self.multi_season:
@@ -60,7 +85,12 @@ class series(object):
 
 
 class episode(object):
-    def __init__(self):
+    def __init__(self, attrs=None):
+        if attrs:
+            for attr, val in attrs.items():
+                setattr(self, attr, val)
+            return
+
         self.fanart = None
         self.thumb = None
         self.title = None
@@ -73,6 +103,9 @@ class episode(object):
         self.drm = None
         self.license_url = None
         self.license_key = None
+
+    def __repr__(self):
+        return type(self).__name__ + '(' + str(dict(vars(self).items())) + ')'
 
     def get_title(self):
         return 'Ep {0} - {1}'.format(self.episode_no, self.episode_name)
@@ -98,7 +131,12 @@ class episode(object):
 
 
 class channel(object):
-    def __init__(self):
+    def __init__(self, attrs=None):
+        if attrs:
+            for attr, val in attrs.items():
+                setattr(self, attr, val)
+            return
+
         self.fanart = None
         self.thumb = None
         self.title = None
@@ -107,6 +145,9 @@ class channel(object):
         self.url = None
         self.id = None
         self.drm = None
+
+    def __repr__(self):
+        return type(self).__name__ + '(' + str(dict(vars(self).items())) + ')'
 
     def get_title(self):
         return '{0} - {1}'.format(self.title, self.desc)
@@ -123,3 +164,70 @@ class channel(object):
         params = urlparse.parse_qsl(url)
         for item in params.keys():
             setattr(self, item, urllib.unquote_plus(params[item]))
+
+
+class CacheObj():
+
+    maxEntries = 50
+    maxTTL = datetime.timedelta(seconds=(60*60*12))
+
+    def __init__(self):
+        self.win = xbmcgui.Window(10000)
+
+    def getData(self, url, headers={},
+                name=None, noCache=False, expiry=None, data=None):
+
+        now = datetime.datetime.now()
+
+        if noCache or not name:
+            return CacheObj.fetch_url(url=url, headers=headers)
+
+        if not data:
+            rawData = self.win.getProperty('%s|%s' % (name, url))
+            if rawData:
+                try:
+                    cachedData = eval(rawData)
+                    if cachedData[0] > now:
+                        return cachedData[1]
+                except Exception as e:
+                    utils.log('Error with eval of cached data: {0}'.format(e))
+                    #utils.log(rawData)
+
+        if not isinstance(expiry, datetime.datetime):
+            expiry = now + CacheObj.maxTTL
+
+        if not data:
+            data = CacheObj.fetch_url(url=url, headers=headers)
+
+        cachedData = (expiry, data)
+        self.win.setProperty('%s|%s' % (name, url), repr(cachedData))
+
+        rawURLs = self.win.getProperty('%s|urls' % name)
+        urls = eval(rawURLs) if rawURLs else deque()
+
+        if url in urls:
+            urls.remove(url)
+        urls.append(url)
+
+        if len(urls) > CacheObj.maxEntries:
+            oldUrl = urls.popleft()
+            self.win.setProperty('%s|%s' % (name, oldUrl), '')
+
+        self.win.setProperty('%s|urls' % name, repr(urls))
+
+        return cachedData[1]
+
+    @staticmethod
+    def fetch_url(url, headers={}):
+        """
+        Use custom session to grab URL and return the text
+        """
+        with session.Session(force_tlsv1=True) as sess:
+            res = sess.get(url, headers=headers)
+            try:
+                data = json.loads(res.text)
+                return data
+            except ValueError as e:
+                utils.log('Error parsing JSON, response is {0}'
+                            .format(res.text))
+                raise e