--- a/resources/lib/config.py
+++ b/resources/lib/config.py
@@ -6,15 +6,17 @@ ISSUE_API_AUTH = 'eGJtY2JvdDo1OTQxNTJjMTBhZGFiNGRlN2M0YWZkZDYwZGQ5NDFkNWY4YmIzOG
 GIST_API_URL = 'https://api.github.com/gists'
 
 CATEGORIES = ['Live TV', 'All Shows']
-GENRES_URL = "https://tv-api.9now.com.au/v1/genres?device=android&take=99999"
-TVSERIES_URL = "https://tv-api.9now.com.au/v1/tv-series?device=android&take=99999"
-TVSERIESQUERY_URL = "https://tv-api.9now.com.au/v1/pages/genres/{0}?device=web&take=99999"
-EPISODEQUERY_URL = "https://tv-api.9now.com.au/v1/pages/tv-series/{0}/seasons/{1}/episodes/{2}?device=android"
-LIVETV_URL = "https://tv-api.9now.com.au/v1/pages/livestreams?device=web"
+GENRES_URL = "https://tv-api.9now.com.au/v2/genres?device=web"
+TVSERIES_URL = "https://tv-api.9now.com.au/v2/tv-series?device=web"
+TVSERIESQUERY_URL = "https://tv-api.9now.com.au/v2/pages/genres/{0}?device=web"
+EPISODEQUERY_URL = "https://tv-api.9now.com.au/v2/pages/tv-series/{0}/seasons/{1}/episodes{2}?device=web"
+LIVETV_URL = "https://tv-api.9now.com.au/v2/pages/livestreams?device=web"
 
 BRIGHTCOVE_URL = "http://c.brightcove.com/services/mobile/streaming/index/master.m3u8?videoId={0}"
 BRIGHTCOVE_DRM_URL = "https://edge.api.brightcove.com/playback/v1/accounts/{0}/videos/ref:{1}"
 BRIGHTCOVE_ACCOUNT = "4460760524001"
 BRIGHTCOVE_KEY = "BCpkADawqM1TWX5yhWjKdzhXnHCmGvnaozGSDICiEFNRv0fs12m6WA2hLxMHM8TGAEM6pv7lhJsdNhiQi76p4IcsT_jmXdtEU-wnfXhOBTx-cGR7guCqVwjyFAtQa75PFF-TmWESuiYaNTzg"
+
+USER_AGENT = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/61.0.3163.100 Safari/537.36'
 
 MAX_HLS_QUAL = 4