--- a/resources/lib/menu.py
+++ b/resources/lib/menu.py
@@ -20,7 +20,7 @@ def list_categories():
         listing = []
         categories = config.CATEGORIES
         for category in categories:
-            li = xbmcgui.ListItem(category)
+            li = xbmcgui.ListItem(category, offscreen=True)
             url_string = '{0}?action=listcategories&category={1}'
             url = url_string.format(_url, category)
             is_folder = True
@@ -29,13 +29,13 @@ def list_categories():
         genres = comm.list_genres()
         for g in genres:
             li = xbmcgui.ListItem(g.title, iconImage=g.thumb,
-                                  thumbnailImage=g.thumb)
+                                  thumbnailImage=g.thumb, offscreen=True)
             li.setArt({'fanart': g.fanart})
             url_string = '{0}?action=listcategories&category=genre&genre={1}'
             url = url_string.format(_url, urllib.quote_plus(g.title))
             is_folder = True
             listing.append((url, li, is_folder))
-        li = xbmcgui.ListItem('Settings')
+        li = xbmcgui.ListItem('Settings', offscreen=True)
         listing.append(('{0}?action=settings'.format(_url), li, is_folder))
         xbmcplugin.addDirectoryItems(_handle, listing, len(listing))
         xbmcplugin.endOfDirectory(_handle)
@@ -51,7 +51,7 @@ def make_episodes_list(url):
         listing = []
         for e in episodes:
             li = xbmcgui.ListItem(e.title, iconImage=e.thumb,
-                                  thumbnailImage=e.thumb)
+                                  thumbnailImage=e.thumb, offscreen=True)
             li.setArt({'fanart': e.fanart})
             url = '{0}?action=listepisodes{1}'.format(_url, e.make_kodi_url())
             is_folder = False
@@ -78,7 +78,7 @@ def make_live_list(url):
         listing = []
         for c in channels:
             li = xbmcgui.ListItem(c.title, iconImage=c.thumb,
-                                  thumbnailImage=c.thumb)
+                                  thumbnailImage=c.thumb, offscreen=True)
             li.setArt({'fanart': c.fanart})
             url = '{0}?action=listchannels{1}'.format(_url, c.make_kodi_url())
             is_folder = False
@@ -109,7 +109,7 @@ def make_series_list(url):
         listing = []
         for s in filtered:
             li = xbmcgui.ListItem(s.title, iconImage=s.thumb,
-                                  thumbnailImage=s.thumb)
+                                  thumbnailImage=s.thumb, offscreen=True)
             li.setArt({'fanart': s.fanart})
             url = '{0}?action=listseries{1}'.format(_url, s.make_kodi_url())
             is_folder = True