… addon.xml
… resources/lib/play.py
… resources/lib/upnext.py