There is probably a better way to do this... but this PR tries to speed up plugin navigation by implementing the following:
- cache 9Now API request JSON response data
- filter shows by primary and secondary genres
- cache processed/filtered data
- offscreen ListItem creation

The cache just stores serialised data as a window property. 50 items are stored for 12 hours.

TODO:

- [ ] method to flush cache
- [ ] method to manually refresh plugin directory listings
- [x] method to request API data in background ~~then compare to cached data~~ to ensure directory listings are always fresh ~~(maybe)~~
- [ ] addon settings for cache (maybe)