@glennguy 2020-08-15
Thanks for the PR @MoojMidge, a lot of stuff here!

Wrt caching - some of our other addons use the [script.common.plugin.cache](https://kodi.wiki/view/Add-on:Common_plugin_cache) module. To be honest though I'm finding myself removing the caching as users have sometimes complained about not having up to date listings, even if the caching was set to 1 hour. API access times used to be an issue with ABC a few years back however their latest version is quite fast.
Have you found the 9Now response times to be laggy?

> Speed up menu creation by creating ListItem offscreen

This ListItem constructor is only available from kodi 18+, it will break for all users still languishing on older versions.

IMO the biggest speedup you can get is from the setting in addon.xml to allow the python interpreter to keep running while still browsing within the same addon, see here: https://github.com/aussieaddons/plugin.video.sbs/blob/master/addon.xml#L11
It works a treat especially on slower devices like RPi

Interested in your thoughts