@MoojMidge 2020-08-15
@glennguy I did some quick testing, just opening the All Series folder ten times for each scenario:

  | Original + reuselanguageinvoker | Original | Cache enabled, offscreen off | Cache enabled, offscreen on | Time Reduction (Original to PR)
-- | -- | -- | -- | -- | --
Average fetch time (s) |   | 00.89 | 00.10 | 00.10 | 88%
Average build list time (s) |   | 11.00 | 11.11 | 00.20 | 98%
Average total time (s) | 06.96 | 11.90 | 11.21 | 00.30 | 97%

Note that the network request appears to be cached anyway, the response header has a TTL of 5 minutes, and so the results above exclude the time taken to originally fetch the data, which takes ~6s on my 100Mbps internet connection.

The averaged cached network request takes 0.89s, but whether or not it actually gets cached, and how long for, will be dependent on an individuals internet/network configuration. Potentially it may never be cached at all, or best case cached for 5 minutes before being requested again.

The biggest speedup appears to be from offscreening the ListItem creation, reducing the time to build the listing from ~11s to 0.20s

If the network request is cached, the local cache reduces the time to fetch the data from 0.89s to 0.10s
If the network request is not cached, the local cache reduces the time to fetch the data from ~6s to 0.10s

If the network request is cached, keeping the Python interpreter open reduces the whole process from 11.9s to 6.96s

> This ListItem constructor is only available from kodi 18+, it will break for all users still languishing on older versions.

Could use a wrapper to create the ListItem, with a try/except to use the offscreen argument, falling back to the original way if it fails?


> Wrt caching - some of our other addons use the [script.common.plugin.cache](https://kodi.wiki/view/Add-on:Common_plugin_cache) module. To be honest though I'm finding myself removing the caching as users have sometimes complained about not having up to date listings, even if the caching was set to 1 hour.

If a cache is used, I think using an existing, tested, addon would normally be a better idea, rather that creating a new one from scratch. However this current approach has the following advantages:
- It caches the network request (easy to do with script.common.plugin.cache)
- It caches the generated listings data, including the instances of the custom classes used (not so easy to do with script.common.plugin.cache)
- It replaces/re-processes the network request data with the generated listings data so the listings don't need to be processed generated all the time (not so easy to do with script.common.plugin.cache). Note this is not that big of a deal as actually creating the ListItems takes the most time.
- Cached data is always kept fresh, as the network data is requested in a separate thread if it is older than 5 minutes. This network data is stored and then used when the listings data is next generated ensuring that the listings are always fresh, without the user ever having to experience the 6s delay to fetch the data (pretty difficult to do with script.common.plugin.cache)

The obvious issues with a custom cache is that it hasn't been widely tested, and there is currently very little error handling.