@glennguy 2020-10-07
Hey @MoojMidge sorry for the huge delay in getting back to you on this.

It all looks good, if you want to wrap the listitem creation like you did for sbs that would be great. Also same issue applies for `setIsFolder` in play.py - that call is new for Kodi 18 and breaks for users of previous versions. I'll merge in after that.