@glennguy 2020-10-08
LGTM, thanks for your efforts!
I'll merge this in as-is, if you want to tackle the stuff on the to-do later just open a new PR
Thanks!